requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
amqtt==0.11.3
numpy>=1.26,<2.0
opencv-python==4.10.0.84
paho-mqtt>=2.1,<3.0
PyYAML>=6.0,<7.0
torch==2.2.2
torchvision==0.17.2
ultralytics==8.3.15
